visualization/pyproject.toml
[tool.poetry]
name = "visualization"
version = "0.1.0"
description = ""
authors = ["Hector Soria <hector.soria@with-madrid.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.5"
vpython = "^7.6.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
